backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
aiomysql==0.2.0
pymysql==1.1.1
alembic==1.13.3
redis==5.1.1
celery==5.4.0
python-dotenv==1.0.1
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2
# use anthropic for clause api
anthropic==0.34.2
PyPDF2==3.0.1
python-docx==1.1.2
httpx==0.27.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.0